requirements.txt
ftfy
rich
python-dotenv
pypdf
llama-index-core
llama-index-embeddings-huggingface
pymilvus